mod-library: make it possible to add existing data to library
This patch changes the API a bit. populateNamed is replaced with static
Library.from() that can is given objects of the same form as before.
"file" key still exists to load and json from url, but it can also be
called with "data"-key that is a reference to some existing object.

In addition, there is a a new "indexDepth" key that can be used to set
how deep a Shelf object is indexed for, defaults to 2