multipage-list: Make changing list source work better
This patch makes bunch od changes to how data loading works both internally and externally. Internally,
bunch of loading list update() logic is now shared between lazy and
non-lazy implementations and list now knows to throw out preloaded
object urls when the list data source changes.

Externally, the old forData() method has been replaced with "src" setter
and getter. Getter simply returns the "src" attribute. Setter writes src
attribute if it is a string, and clears it otherwise - but in addition,
the setter calls setSource() method with the argument if it was an array - or first calls fetch(argument)
and then calls setSource with the result of the fetch.